requirements.txt
cyglfw3==3.1.0.2
gl-utils==3.0.0
numpy==1.13.3
Pillow==4.3.0
pyassimp==3.3
PyOpenGL==3.1.1a1